Add files via upload
This project showcases the use of the central limit theorem to find if a treatment is effective on a group of patients.
We have two samples, one sample has experienced the treatment, and the other sample got a placebo treatment.
We find the point estimate, critical value, and standard error. These help us find the confidence interval for the experiment.